new file:   app/LoginBtn.js 	modified:   app/layout.js 	modified:   app/list/DetailLink.js 	modified:   app/list/page.js 	modified:   app/list2/page.js 	modified:   next.config.mjs 	modified:   package-lock.json 	modified:   package.json 	new file:   pages/api/auth/[...nextauth].js 	modified:   pages/api/delete.js 	modified:   pages/api/post/new.js 	modified:   util/database.js

+++ pages/api/delete.js
@@ -1,19 +1,29 @@
 import { connectDB } from "@/util/database";
 import { ObjectId } from "mongodb";
+import { getServerSession } from "next-auth";
+import { authOptions } from "./auth/[...nextauth]";
 
 async function handler (req, res) {
-
+    
+    
     if(req.method==="POST"){
-        try{
-            const db = ( await connectDB ).db("forum");
-            await db.collection("post").deleteOne({_id: new ObjectId(req.body)})
-            
-            return res.status(200).json("삭제 완료")
+
+        const session = await getServerSession(req, res, authOptions);
+        const db = ( await connectDB ).db('forum')
+        const result = await db.collection('post').findOne({_id: new ObjectId( req.body)})
+
+        if(result.writer === session.user.email){
+
+            await db.collection('post').deleteOne({_id: new ObjectId(req.body)})
+
+            console.log("삭제 완료")
+
+            return res.status(200).redirect("/list");
         }
-        catch (error) {
-            return res.status(500).json("삭제 살패")
+        else{
+            console.log("삭제 불가")
+            return res.status(200).redirect("/list");
         }
-        
     }
 }
